+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Release 2.4.2 (in development)
 
+### Features added
+- [#2049](https://github.com/scality/metalk8s/issues/2049) - Deploy
+[prometheus-adapter](https://github.com/DirectXMan12/k8s-prometheus-adapter/)
+to implement the `metrics.k8s.io` API, to support `kubectl top` and other
+consumers of this API
+(PR [#2057](https://github.com/scality/metalk8s/pull/2057))
+
 ## Release 2.4.1
 
 ### Features added